ai-service/ requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
requests==2.31.0
pillow==10.2.0
